refactor : 자잘한 오류 수정 및 피드백 반영
- CustomExceptionHandler ErrorResponse 타입 지정
- 점수 갱신 updateScore -> increaseScore 네이밍 수정
- Top3 랭킹 조회 시, limit(3) 반영
- newRank 갱신 시, for문 개선

+++ src/main/java/com/gamzabat/algohub/feature/group/ranking/dto/GetRankingResponse.java
@@ -9,10 +9,10 @@ public class GetRankingResponse {
 	private String userNickname;
 	private String profileImage;
 	private Integer rank;
-	private Long solvedCount;
+	private Integer solvedCount;
 	private String rankDiff;
 
-	public GetRankingResponse(String userNickname, String profileImage, Integer rank, Long solvedCount,
+	public GetRankingResponse(String userNickname, String profileImage, Integer rank, Integer solvedCount,
 		String rankDiff) {
 		this.userNickname = userNickname;
 		this.profileImage = profileImage;
